requirements.txt
python-telegram-bot==21.6
youtube-transcript-api
requests